# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: StellarMind API
  description: REST API for the StellarMind agent orchestration platform
  version: '1.0.0'
  contact:
    name: StellarMind Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
paths:
...
  /orchestration/preview:
    post:
      summary: Preview an orchestration plan and quote work before execution
      operationId: previewOrchestration
      tags: [Orchestration]
      description: >
        Generates and validates an orchestration plan, returning validated steps, per-step quotes,
        total quote, expiry timestamp, and any planner-provider usage. Makes zero premium agent calls
        or settlement attempts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchestrationInput'
      responses:
        '200':
          description: Plan preview with quotes and expiry
          content:
            application/json:
              schema:
                type: object
                properties:
                  planId:
                    type: string
                  task:
                    type: string
                  budget:
                    type: number
                  plan:
                    type: string
                  steps:
                    type: array
                    items:
                      type: object
                  totalQuote:
                    type: string
                  expiresAt:
                    type: string
                  usage:
                    type: object
        '400':
          description: Validation error
...
  /runs/{runId}:
    get:
      summary: Get run detail
      operationId: getRunDetail
      tags: [Orchestration]
      description: >
        Retrieve the full detail of a run, including its current status, output (when completed),
        and error information (when failed). This is the endpoint referred to by the Location header
        in 202 responses.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '404':
          description: Run not found
  /runs:
    get:
      summary: List recent runs
      operationId: listRuns
      tags: [Orchestration]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Recent runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunDetail'
...
components:
  schemas:
...
    OrchestrationInput:
      type: object
      required: [task]
      properties:
        task:
          type: string
          description: The task description for the orchestrator to plan and execute
        budget:
          type: number
          description: Maximum USDC budget for this orchestration run
        mode:
          type: string
          enum: [sync, async]
          default: sync
          description: Submission mode. "async" returns 202 immediately.
        idempotencyKey:
          type: string
          maxLength: 256
          description: >
            Optional idempotency key (alternative to the Idempotency-Key header). Prevents duplicate
            runs when the same key is used.
...
    RunDetail:
      type: object
      description: |
        Full run record including status, output, and error information.
        - running: `output` is `null` and `outputAvailable` is `false`
        - failed: `output` is `null`, `outputAvailable` is `false`, and `error` holds the failure data
        - completed: `output` contains the persisted orchestration results and `outputAvailable` is `true`
        - legacy records without persisted output: `output` remains `null` and `outputAvailable` is `false`
      properties:
        runId:
          type: string
        status:
          type: string
          enum: [running, completed, failed]
        task:
          type: string
        budget:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        outputAvailable:
          type: boolean
          description: true only when the final output has been persisted and is safe to read
        output:
          type: array
          items:
            type: object
          nullable: true
          description: Persisted orchestration results. Null while running, after failures, or for legacy records without stored output.
        error:
          type: object
          nullable: true
          description: Error details (null unless run failed)
          properties:
            message:
              type: string
            code:
              type: string
        summary:
          type: object
          nullable: true
        events:
          type: array
          items:
            type: object
...